Clean up scrolling delegate code for 2D (#122651)
Clean up scrolling delegate code for 2D
